+++ src/main/java/com/ataberkcanitez/migroscouriertracking/infra/adapters/courier/websocket/CourierLocationSubject.java
@@ -6,5 +6,5 @@ import com.ataberkcanitez.migroscouriertracking.domain.location.model.Location;
 public interface CourierLocationSubject {
     void attachObserver(CourierLocationSubscriber observer);
     void detachObserver(CourierLocationSubscriber observer);
-    void notifyObservers(int courierId, Location location);
+    void notifyObservers(long courierId, Location location);
 }
